dom/media/mediasource/test/test_SeekNoData_mp4.html

<!DOCTYPE HTML>
<html>
<head>
  <title>MSE: basic functionality</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="mediasource.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

// Avoid making trouble for people who fix rounding bugs.
function fuzzyEquals(a, b) {
  return Math.abs(a - b) < 0.01;
}

runWithMSE(function(ms, el) {
  el.controls = true;
  once(ms, "sourceopen").then(function() {
    ok(true, "Receive a sourceopen event");
    const audiosb = ms.addSourceBuffer("audio/mp4");
    const videosb = ms.addSourceBuffer("video/mp4");
    el.addEventListener("error", function(e) {
      ok(false, "should not fire '" + e + "' event");
    });
    is(el.readyState, el.HAVE_NOTHING, "readyState is HAVE_NOTHING");
    try {
      el.currentTime = 3;
    } catch (e) {
      ok(false, "should not throw '" + e + "' exception");
    }
    is(el.currentTime, 3, "currentTime is default playback start position");
    is(el.seeking, false, "seek not started with HAVE_NOTHING");
    const metadataPromises = [];
    metadataPromises.push(fetchAndLoad(audiosb, "bipbop/bipbop_audio", [ "init" ], ".mp4"));
    metadataPromises.push(fetchAndLoad(videosb, "bipbop/bipbop_video", [ "init" ], ".mp4"));
    metadataPromises.push(once(el, "loadedmetadata"));
    Promise.all(metadataPromises)
    .then(function() {
      const p = once(el, "seeking");
      el.play();
      el.currentTime = 5;
      is(el.readyState, el.HAVE_METADATA, "readyState is HAVE_METADATA");
      is(el.seeking, true, "seek not started with HAVE_METADATA");
      is(el.currentTime, 5, "currentTime is seek position");
      return p;
    })
    .then(function() {
      ok(true, "Got seeking event");
      const seekedPromises = [];
      seekedPromises.push(once(el, "seeked"));
      seekedPromises.push(fetchAndLoad(audiosb, "bipbop/bipbop_audio", range(5, 9), ".m4s"));
      seekedPromises.push(fetchAndLoad(videosb, "bipbop/bipbop_video", range(6, 10), ".m4s"));
      return Promise.all(seekedPromises);
    })
    .then(function() {
      ok(true, "Got seeked event");
      ok(el.currentTime >= 5, "Time >= 5");
      once(el, "ended").then(SimpleTest.finish.bind(SimpleTest));
      ms.endOfStream();
    });
  });
});

</script>
</pre>
</body>
</html>
